Rename header.
