+++ rosserial_embeddedlinux/src/ros_lib/embedded_linux_hardware.h
